fix(zap_dev): Wrap interpolated identifiers in braces
When component variables are interpolated in strings, the replacement identifier can cause invalidate Dart code to be generated if the identifier is not wrapped in braces already.

For example:

```html
<script>
  String? value;
  String? interpolated;

  $: interpolated = '$value';
</script>
```

Here, the interpolated string currently becomes something like `$_$v2 /* value */` which is invalid Dart code, when we really want `${_$v2 /* value */}`.